ARM64: Fix register width for disassembly of normal memory access instructions
For the normal memory access instructions, the width of Rd and Rm could be 32
or 64 bits, depending on the size, opc and option of each instruction, while
the base register is always 64 bit width. The origin GEM5 does not give a
correct disassembly result. This patch fixes the problem.

The instructions affected inlucde:
- Load/Store(immediate)
- Load/Store(pre-index)
- Load/Store(post-index)
- Load/Store(register)
- PRFM (register)
- LDR(literal)
- LDRSW(literal)
- PRFM (immediate)
- PRFM (literal)

Change-Id: Ie6f58a29712a91d0d9b4923b52a3ab84dddf4f2d
Signed-off-by: Ian Jiang <ianjiang.ict@gmail.com>
Signed-off-by: Lv Zheng <zhenglv@hotmail.com>