RISCV: Fix disassembling of immediate for c.lui
For compressed instruction c.lui, the 6-bit immediate is left-shifted by 12 bits
in decoding. While the original Gem5 gives the left-shifted value directly in
disassembly. This patch fixes the problem by adding a new format CILuiOp and a
new template CILuiExecute to resume the immediate before outputting it in
disassembly.
Note: The immediate is sign-extended to 20-bit to be compatible with GCC.

Change-Id: If73f72d3e8f85a8b10ce7a323379d8ad6c4c3085
Signed-off-by: Ian Jiang <ianjiang.ict@gmail.com>