base,cpu: Add named branch target support in SymbolTable.
This patch adds a mechanism to enhance branch diassembly support.

GEM5 disassemblies use a special format for branch targets:
 bl #address to bl symbol
 b.ne <symbol+offset> to b.ne #offset	// <symbol+offset>
And the latter format cannot be re-compiled into objects by GNU assembler.

This patch adds a mechnism in the symtab class to allow branch
instructions to mark such offset branch targets so that when GEM5 dumps
disassemblies, matched targets can be converted into a string label in the
format of "symbol_offset".
This feature is only enabled for BaseSimpleCPU when TRACING_ON is enabled
which is the only mode used in the follow-up patches to dump checkpoint
assemblies.

Noted that there is no user of markTarget() introduced in this patch. The
feature will be used by further commits.

Change-Id: Idc1a83d7c52b0b9dae867cdc03f0329f141e1dc0
Signed-off-by: Lv Zheng <zhenglv@hotmail.com>